+++ Togather/Togather/Sources/Scene/Login/View/LoginView.swift
@@ -1,0 +1,51 @@
+import SwiftUI
+
+struct LoginView: View {
+    @State var email = ""
+    @State var password = ""
+    
+    @StateObject var loginViewModel = LoginViewModel()
+
+    var body: some View {
+        GeometryReader { proxy in
+            ColorManager.BackgroundColor.ignoresSafeArea()
+            ZStack {
+                NavigationLink(destination: TabBarView(), tag: 1, selection: $loginViewModel.viewTag ) { EmptyView() }
+                VStack(alignment: .leading, spacing: 12) {
+                    Spacer()
+                    LoginTitle()
+                    
+                    LoginTextField(
+                        id: $loginViewModel.email,
+                        password: $loginViewModel.password
+                    )
+                    
+                    AuthButton(
+                        isLoading: $loginViewModel.showProgrees,
+                        text: "로그인",
+                        action: {
+                            loginViewModel.showProgrees = true
+                            loginViewModel.login()
+                        }
+                    )
+                    
+                    LoginNavigationButton()
+                    Spacer()
+                    LoginImageButton()
+                }
+            }
+            .padding([.horizontal, .bottom], 16)
+            .alert("안내", isPresented: $loginViewModel.showError) {
+                Button("확인", role: .cancel) { }
+            } message: {
+                Text(loginViewModel.errorMessage)
+            }
+        }.navigationBarHidden(true)
+    }
+}
+
+struct SignInView_Previews: PreviewProvider {
+    static var previews: some View {
+        LoginView()
+    }
+}
